requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.4.0
beautifulsoup4==4.11.1
cattrs==1.10.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
cssselect==1.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.45.0
django-bootstrap-breadcrumbs==0.9.2
django-ckeditor==6.4.2
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-social-share==2.2.1
django-taggit==3.0.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==8.11.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.11.4
lxml==4.9.0
Markdown==3.3.7
markdown2==2.4.3
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.3.1
psycopg2==2.9.1
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.1.0
pyquery==1.4.3
python-dateutil==2.8.2
python-memcached==1.59
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2021.1
readtime==2.0.0
requests==2.26.0
requests-cache==0.9.4
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
text-unidecode==1.3
tzdata==2022.1
url-normalize==1.4.3
urllib3==1.26.6
whitenoise==5.3.0
zipp==3.8.0
